feat(): 0715 신고 삭제 필터(정지 때 내역삭제 불가)/신고하기/정지,정지해제 ex)다른테이블 연결되면 내 신고조회/신고하기 수정하기 SW

+++ src/main/java/com/momo/momopjt/report/ReportServiceImpl.java
@@ -8,6 +8,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -16,17 +17,17 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Transactional
 @Log4j2
-public class ReportServiceImpl implements ReportService{
+public class ReportServiceImpl implements ReportService {
   private final ReportRepository reportRepository;
   private final ModelMapper modelMapper;
   private final UserRepository userRepository;
 
-  //생성
-  @Override
-  public void addReport(Report report) {
-    //Report 를 받아와서 db에 저장한다.
+  //신고하기 (추가)
+  public void addReport(ReportDTO reportDTO) {
+    Report report = modelMapper.map(reportDTO, Report.class);
     reportRepository.save(report);
   }
+
   //자신 유저 아이디로 조회
   @Override
   public List<ReportDTO> readReport(User reporterNo) {
@@ -45,27 +46,77 @@ public class ReportServiceImpl implements ReportService{
         .map(report -> modelMapper.map(report, ReportDTO.class))
         .collect(Collectors.toList());
   }
+
   //유저 제제 (수정)
   @Override
-  public void updateReport(ReportDTO reportDTO) {
+  public void killReport(ReportDTO reportDTO) {
+    log.info("...... [updateReport START]..........KSW");
+
     Optional<Report> justice = reportRepository.findById(reportDTO.getReportNo());
     if (justice.isPresent()) {
+      log.info("...... [jusstice present]..........KSW");
       Report report = justice.get();
       // reportResult 가 1일 경우 2로 변경하여 저장
       if (report.getReportResult() == '1') {
+        log.info("...... [reportResult == 1]..........KSW");
         report.setReportResult('2');
-        reportRepository.save(report);
-//        userRepository.findById()
-      }
-    } else {
-      if (reportDTO.getReportResult() != '1') {
-        log.info("..........[이미 싸늘해진 시체다]..........KSW");
+        // 사용자의 userState 값 변경
+        Optional<User> userOptional = userRepository.findById(report.getReportedNo().getUserNo());
+        if (userOptional.isPresent()) {
+          log.info("...... [userOptional is Present !!]..........KSW");
+          User user = userOptional.get();
+          user.setUserState('2');
+          userRepository.save(user);
+        } else {
+          if (reportDTO.getReportResult() != '1') {
+            log.info("..........[이미 싸늘해진 시체다]..........KSW");
+          }
+        }
       }
     }
   }
-// 삭제
-@Override
-public void deleteReport(Long reportNo) {
-  reportRepository.deleteById(reportNo);
+  //유저 제제 해제(수정) 제제랑 숫자만 다름
+  @Override
+  public void safeReport(ReportDTO reportDTO) {
+    log.info("...... [updatReport START]..........KSW");
+    Optional<Report> sage = reportRepository.findById(reportDTO.getReportNo());
+    if (sage.isPresent()) {
+      log.info("...... [sage present]..........KSW");
+      Report report = sage.get();
+      // reportResult 가 2일 경우 1로 변경하여 저장
+      if (report.getReportResult() == '2') {
+        log.info("...... [reporResult == 2]..........KSW");
+        report.setReportResult('1');
+        // 사용자의 userState 값 변경
+        Optional<User> userOptional = userRepository.findById(report.getReportedNo().getUserNo());
+        if (userOptional.isPresent()) {
+          log.info("...... [userOptional is Present? !!]..........KSW");
+          User user = userOptional.get();
+          user.setUserState('1');
+          userRepository.save(user);
+        } else {
+          if (reportDTO.getReportResult() != '2') {
+            log.info("..........[살아남남]..........KSW");
+          }
+        }
+      }
+    }
+  }
+
+  // 삭제
+  @Override
+  public void deleteReport(Long reportNo) {
+    Optional<Report> optionalReport = reportRepository.findById(reportNo);
+    if (optionalReport.isPresent()) {
+      Report report = optionalReport.get();
+      if (report.getReportResult() == '1') {
+        reportRepository.deleteById(reportNo);
+      } else {
+        System.out.println("회원 상태가 이상하여 삭제처리가 어렵습니다");
+      }
+    } else {
+      System.out.println("회원을 찾을 수 없습니다!");
+    }
+  }
 }
-}
+
